App should now properly intercept all nfc tags.
