+++ app/build.gradle
@@ -13,8 +13,8 @@ android {
         applicationId "com.TrakEngineering.FluidSecureHubFOBapp"
         minSdkVersion 15
         targetSdkVersion 25
-        versionCode 1
-        versionName "0.1.0"
+        versionCode 2
+        versionName "0.2.0"
         multiDexEnabled true
     }
     buildTypes {
